pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iseeyou-baseline"
version = "0.1.0"
description = "I SEE YOU capstone baseline for real/generated/deepfake face classification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "timm>=1.0.0",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "PyYAML>=6.0.0",
  "tqdm>=4.66.0",
  "facenet-pytorch>=2.6.0",
  "yt-dlp>=2024.8.6"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["iseeyou*"]
